requirements.txt
streamlit==1.32.0
boto3==1.34.50
pandas==2.2.1
numpy==1.26.4
Pillow==10.2.0
opencv-python==4.9.0.80
pdf2image==1.17.0
PyMuPDF==1.23.26
requests==2.31.0
python-dotenv==1.0.1
matplotlib==3.8.3 